Titouan

+++ src/main/java/com/example/geodevineur/dep_reg/DepReg.java
@@ -38,7 +38,7 @@ public abstract class DepReg {
         return population;
     }
 
-    public int getSurface() {
+    public double getSurface() {
         return surface;
     }
